Fixes and Tweaks
- Further tweaks of project settings for performance optimization.
- Fixes to half bread's commit, not many: some ambiguities of "Debug", one "ermh, you forgot the f after the number", and one currently unused script that still used double variables).

+++ Assets/[Dev5]Environment/TerrainBehavior.cs
@@ -7,16 +7,16 @@ public class TerrainBehavior : MonoBehaviour
     #region Variables
 
     #region Terrain Properties
-    public double Area;
-    public double Temperature;
-    public double Moles;
-    public double MolarWeight = 0.0401;
-    public double MassPerSquareMeter;
-    public double Albedo;
-    public double Emissivity;
-    public double HeatCapacity;
-    public double HeatInertia;
-    public double LocalLatitude;
+    public float Area;
+    public float Temperature;
+    public float Moles;
+    public float MolarWeight = 0.0401f;
+    public float MassPerSquareMeter;
+    public float Albedo;
+    public float Emissivity;
+    public float HeatCapacity;
+    public float HeatInertia;
+    public float LocalLatitude;
     #endregion
 
     #region Reference Scripts
